t extract approvaltests tests to separate module
