+++ approvaltests-tests/src/test/java/org/approvaltests/JunitUpgrade.java
@@ -1,4 +1,4 @@
-package com.spun;
+package org.approvaltests;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
