+++ pom.xml
@@ -94,6 +94,7 @@
         <module>approvaltests-util</module>
         <module>approvaltests-util-tests</module>
         <module>approvaltests</module>
+        <module>approvaltests-tests</module>
         <module>approvaltests-hadoop</module>
         <module>counter_display</module>
         <module>html_locker</module>
