+++ run_tests_compile_dependencies.sh
@@ -1,0 +1,2 @@
+mvn install -DskipTests
+mvn verify -pl approvaltests-tests/
