+++ run_tests_without_compile_dependencies.sh
@@ -1,0 +1,1 @@
+mvn verify -pl approvaltests-tests/
