--- run_utils_test_compile_dependencies.sh
@@ -1,2 +1,0 @@
-mvn install -DskipTests
-mvn verify -pl approvaltests-util-tests/
