--- run_utils_test_without_compile_dependencies.sh
@@ -1,1 +1,0 @@
-mvn verify -pl approvaltests-util-tests/
